TripExpensesProject/IAM/requirements.txt
django
djangorestframework
pyjwt
pymysql
mysqlclient